fixed multiple reports folders
